requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
biopython==1.85
bleach==6.1.0
blinker==1.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
datasets==3.5.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.7.0
docker-pycreds==0.4.0
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.9.0
flash_attn==2.7.4.post1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
httplib2==0.20.2
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-archive==3.4.0
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning-utilities==0.14.3
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.24.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
omegaconf==2.3.0
opt_einsum==3.4.0
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pycparser==2.21
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tokenizers==0.21.1
tomli==2.0.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchmetrics==1.7.1
torchvision==0.20.1+cu124
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
transformers==4.51.0
triton==3.1.0
types-python-dateutil==2.8.19.14
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.13
wadllib==1.3.6
wandb==0.19.9
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xxhash==3.5.0
yarl==1.19.0
zipp==1.0.0